RM-804: Should a user granted Site Manager role be able to delete the Records Management site?
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/modules/recordsmanagement/HEAD@56405 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ rm-server/test/java/org/alfresco/module/org_alfresco_module_rm/test/util/BaseRMTestCase.java
@@ -406,20 +406,29 @@ public abstract class BaseRMTestCase extends RetryingTransactionHelperTestCase
         filter.disableBehaviour();
         try
         {
-            Set<NodeRef> holds = freezeService.getHolds(filePlan);
-            for (NodeRef hold : holds)
+            if (nodeService.exists(filePlan) == true)
             {
-                freezeService.relinquish(hold);
+                Set<NodeRef> holds = freezeService.getHolds(filePlan);
+                for (NodeRef hold : holds)
+                {
+                    freezeService.relinquish(hold);
+                }
             }
             
-            // Delete the folder
-            nodeService.deleteNode(folder);
-    
-            // Delete the site
-            siteService.deleteSite(siteId);
+            if (nodeService.exists(folder) == true)
+            {
+                // Delete the folder
+                nodeService.deleteNode(folder);
+            }
+            
+            if (siteService.getSite(siteId) != null)
+            {
+                // Delete the site
+                siteService.deleteSite(siteId);
+            }
     
             // delete the collaboration site (if required)
-            if (isCollaborationSiteTest() == true)
+            if (isCollaborationSiteTest() == true && siteService.getSite(COLLABORATION_SITE_ID) != null)
             {
                 siteService.deleteSite(COLLABORATION_SITE_ID);
             }
